Oliver (Added functionality to basket and order.)

+++ app/Http/Controllers/OrderController.php
@@ -8,34 +8,70 @@ use Illuminate\Support\Facades\Auth;
 use App\Models\cart;
 use App\Models\Order;
 use App\Models\Book;
+use App\Models\Payment;
 
 class OrderController extends Controller
 {
-    public function index() {
-        $user_id = Auth::id();
-        $basket = cart::where('user_id', $user_id)->get();
+    public function index(Request $request) {
         $books = collect();
+        if (Auth::check()) {
+            $user_id = Auth::id();
+            $basket = cart::where('user_id', $user_id)->get();
+            
+        } else {
+            if ($request->session()->has('books')) {
+                $basket = $request->session()->get('books');
+            } else {
+                $basket = collect();
+            }
+        }
         //$amounts = array();
+
         foreach ($basket as $elem) {
-            $books->push(Book::where('id', $elem['book_id'])->get());
+            if (Auth::check()) {
+                $books->push(Book::where('id', $elem['book_id'])->get());
+            } else {
+                $books->push(Book::where('id', $elem)->get());
+            }
             //array_push($amounts, $elem['Amount']);     
         }
-        return view('order', [
-            'books' => $books
+        return view('/order', [
+            'books' => $books,
             //'amounts' => $amounts,
         ]);
     }
 
-    public function create() {
-        $user_id = Auth::id();
-        $books = cart::where('user_id', $user_id)->get('book_id');
+    public function create(Request $request) {
+        if (Auth::check()) {
+            $user_id = Auth::id();
+            $books = cart::where('user_id', $user_id)->get('book_id');
+            $payment = Payment::where('credit_card_no', request('credit_card_no'))->get();
+            if (count($payment) == 0) {
+                $payment = new Payment;
+                $payment->credit_card_no = request('credit_card_no');
+                $payment->user_id = $user_id;
+                $payment->save();
+            }
+        } else {
+            $books = $request->session()->get('books');
+        }
+        
+        
+        //$order_status = new Order_Status;
+        
         foreach($books as $book) {
-            $book = $book['book_id'];
+            if (Auth::check()) {
+                $book = $book['book_id'];
+            }
             $order = new Order;
             $order->book_id = $book;
             $order->save();
         }
-        cart::where('user_id', $user_id)->delete();
+        if (Auth::check()) {
+            cart::where('user_id', $user_id)->delete();
+        } else {
+            $request->session()->put('books', collect());
+        }
         return redirect('basket')->with('success', 'Order Complete');
     }
 }
